# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Juarez Barber API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /barbershops/{barbershopId}/services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services for a barbershop
      parameters:
        - name: barbershopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      security:
        - bearerAuth: []
      summary: Create a service
      parameters:
        - name: barbershopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      security:
        - bearerAuth: []
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      security:
        - bearerAuth: []
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Appointments
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      security:
        - bearerAuth: []
      summary: List sales
      parameters:
        - name: barbershopId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      security:
        - bearerAuth: []
      summary: Create a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    delete:
      operationId: deleteSale
      tags: [sales]
      security:
        - bearerAuth: []
      summary: Delete a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Expenses
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      security:
        - bearerAuth: []
      summary: List expenses
      parameters:
        - name: barbershopId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      security:
        - bearerAuth: []
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      security:
        - bearerAuth: []
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Commissions
...
  /reports/financial-summary:
    get:
      operationId: getFinancialSummary
      tags: [reports]
      security:
        - bearerAuth: []
      summary: Get financial summary (revenue, expenses, profit, commissions)
      parameters:
        - name: barbershopId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"
  /reports/monthly-cashflow:
    get:
      operationId: getMonthlyCashflow
      tags: [reports]
      security:
        - bearerAuth: []
      summary: Get monthly cashflow data for charts
      parameters:
        - name: barbershopId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyCashflow"
  /reports/dashboard:
    get:
      operationId: getDashboard
      tags: [reports]
      security:
        - bearerAuth: []
      summary: Get admin dashboard stats (appointments today, revenue this week, top barbers)
      parameters:
        - name: barbershopId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /reports/appointments-by-status:
    get:
      operationId: getAppointmentsByStatus
      tags: [reports]
      security:
        - bearerAuth: []
      summary: Get appointment counts grouped by status
      parameters:
        - name: barbershopId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentsByStatus"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Service:
      type: object
      required: [id, name, price, durationMinutes, barbershopId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationMinutes:
          type: integer
        barbershopId:
          type: integer
        createdAt:
          type: string
          format: date-time
    ServiceInput:
      type: object
      required: [name, price, durationMinutes]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
...
    Sale:
      type: object
      required: [id, amount, description, barbershopId, barberId, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        description:
          type: string
        barbershopId:
          type: integer
        barberId:
          type: integer
        barberName:
          type: ["string", "null"]
        appointmentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    SaleInput:
      type: object
      required: [amount, description, barbershopId, barberId]
      properties:
        amount:
          type: number
        description:
          type: string
        barbershopId:
          type: integer
        barberId:
          type: integer
        appointmentId:
          type: integer
    Expense:
      type: object
      required: [id, amount, description, category, barbershopId, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        description:
          type: string
        category:
          type: string
        barbershopId:
          type: integer
        createdAt:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [amount, description, category, barbershopId]
      properties:
        amount:
          type: number
        description:
          type: string
        category:
          type: string
        barbershopId:
          type: integer
...
    FinancialSummary:
      type: object
      required: [totalRevenue, totalExpenses, netProfit, totalCommissions, appointmentsCompleted]
      properties:
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        totalCommissions:
          type: number
        appointmentsCompleted:
          type: integer
    MonthlyCashflow:
      type: object
      required: [month, revenue, expenses, profit]
      properties:
        month:
          type: string
        revenue:
          type: number
        expenses:
          type: number
        profit:
          type: number
    DashboardStats:
      type: object
      required: [appointmentsToday, appointmentsThisWeek, revenueToday, revenueThisWeek, pendingAppointments, topBarbers]
      properties:
        appointmentsToday:
          type: integer
        appointmentsThisWeek:
          type: integer
        revenueToday:
          type: number
        revenueThisWeek:
          type: number
        pendingAppointments:
          type: integer
        topBarbers:
          type: array
          items:
            $ref: "#/components/schemas/TopBarber"
    TopBarber:
      type: object
      required: [barberId, name, appointmentsCount, revenue]
      properties:
        barberId:
          type: integer
        name:
          type: string
        appointmentsCount:
          type: integer
        revenue:
          type: number
    AppointmentsByStatus:
      type: object
      required: [pending, confirmed, completed, cancelled]
      properties:
        pending:
          type: integer
        confirmed:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer