# lib/api-zod/src/generated/types/appointmentsByStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Juarez Barber API
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentsByStatus {
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
}