# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Juarez Barber API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  amount: number;
  description: string;
  category: string;
  barbershopId: number;
  createdAt: Date;
}