# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Juarez Barber API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  durationMinutes: number;
  barbershopId: number;
  createdAt: Date;
}